Resolve race condition in jpathAppRunner
Resolve race condition in jpathAppRunner

During unwrap output, there was this error that was occasionally coming up that mentioned the wait group was negative. Upon running some manual tests, it was found that the channel sometimes had empty lines in it. Not sure what causes those empty lines to be emitted to the channel, but ensuring we are decrementing the wait group only on non-empty channel inputs ensures that we are not running into negative wg counts